Identity smart contract added
The Identity smart contract is used to verify the identity of users on the Ethereum blockchain. It has the following functions:

    addIdentity: This function is used to add a new user's identity to the blockchain. Only the manager is able to access this function.

    updateIdentity: This function is used to update a user's identity on the blockchain. Only the manager is able to access this function.

    verifyIdentity: This function is used to verify the identity of a user on the blockchain. Only the manager is able to access this function. The identity is verified using a hash that is generated using the user's address, name, and government ID.

    setManager: This function is used to set the manager who will have access to all the functions of the contract.

The contract also has three events: IdentityAdded, IdentityUpdated, and IdentityVerified, which are triggered when a new identity is added, an existing identity is updated, and when an identity is verified, respectively.